Generalize compose dropdown for re-use
